requirements.txt
# Feel free to add your Python dependencies below. For formatting guidelines, see:
# https://pip.pypa.io/en/latest/reference/requirements-file-format/

apify < 3.0
apify-client
python-dotenv
pydantic >= 2.0
pydantic-ai
